fix: critical QA fixes — metallib name, data race, row index types
Critical fixes from QA review:

1. Fix metallib filename mismatch: hist_updater.mm looked for
   "xgboost_metal.metallib" but CMake produces "xgboost.metallib".
   Pre-compiled Metal library now loads correctly.

2. Fix data race in embedded kernel's global histogram flush: replace
   non-atomic read of hist_out[i] with atomic_load_explicit before
   the CAS loop. Prevents undefined behavior under threadgroup
   contention.

3. Fix row index type mismatch: pre-compiled kernels (histogram.metal,
   partition.metal) used uint (32-bit) for row indices but the host
   stores size_t (64-bit on ARM64). Changed all kernel signatures to
   use ulong for row index parameters.

Also:
- Remove dead eval_splits_pipeline_ field and destructor code
- 29/29 tests pass

+++ plugin/metal/kernels/histogram.metal
@@ -137,7 +137,7 @@ kernel void clear_histogram(
 kernel void build_histogram(
     device const float2*   gpair        [[buffer(0)]],
     device const uint8_t*  gidx         [[buffer(1)]],
-    device const uint*     ridx         [[buffer(2)]],
+    device const ulong*    ridx         [[buffer(2)]],
     device float*          hist         [[buffer(3)]],
     constant uint&         n_rows       [[buffer(4)]],
     constant uint&         row_stride   [[buffer(5)]],
@@ -236,7 +236,7 @@ kernel void build_histogram(
 kernel void build_histogram_dense(
     device const float2*   gpair       [[buffer(0)]],
     device const uint8_t*  gidx        [[buffer(1)]],
-    device const uint*     ridx        [[buffer(2)]],
+    device const ulong*    ridx        [[buffer(2)]],
     device float*          hist        [[buffer(3)]],
     device const uint*     offsets     [[buffer(4)]],
     constant uint&         n_rows      [[buffer(5)]],
